Moving input handling to new class and handle it per stroke
