+++ src/util/XournalTypeList.h
@@ -257,9 +257,7 @@ XOJ_DECLARE_TYPE(PopplerGlibPage, 245);
 XOJ_DECLARE_TYPE(PopplerGlibPageBookmarkIterator, 246);
 XOJ_DECLARE_TYPE(PopplerGlibAction, 247);
 XOJ_DECLARE_TYPE(XojCairoPdfExport, 248);
-XOJ_DECLARE_TYPE(BaseInputDevice, 249);
 XOJ_DECLARE_TYPE(AbstractInputDevice, 250);
-XOJ_DECLARE_TYPE(DirectAxisInputDevice, 251);
 XOJ_DECLARE_TYPE(NewGtkInputDevice, 252);
 XOJ_DECLARE_TYPE(InputSequence, 253);
